% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example_app.R
\name{run_example_app}
\alias{run_example_app}
\title{run_example_app(example)}
\usage{
run_example_app(example)
}
\arguments{
\item{example}{a string title of the app you want to run in the inst/shiny-examples dir}
}
\description{
run our example app; note you must set up your rprofile to run.
}
