% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_filter_PAW.R
\name{s3_filter_PAW}
\alias{s3_filter_PAW}
\title{s3_filter_PAW(bucketlist)}
\usage{
s3_filter_PAW(
  bucketlist = NULL,
  category = "MER",
  subcategory = NULL,
  metadata = FALSE,
  country = NULL
)
}
\arguments{
\item{bucketlist}{list of paths from your aws bucket}

\item{category}{Filter to a specific category "MER", "Financial", or "Narratives".}

\item{subcategory}{Filter to a specific subcategory "Site_Recent", "Site_Historic", "PSNU_Recent","PSNU_Historic", "OU_Recent", or "OU_Historic".}

\item{metadata}{metadata files will be included when set to TRUE}

\item{country}{Filter to a specific country's data.}
}
\description{
Standard function for filtering PAW data in S3, returns a filtered list of bucket items.
}
