% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_read_PAW.R
\name{s3_read_PAW}
\alias{s3_read_PAW}
\title{s3_read_PAW(bucket, path)}
\usage{
s3_read_PAW(bucket, path)
}
\arguments{
\item{bucket}{your aws bucket}

\item{path}{your list of paths to read}
}
\description{
Standard function for reading from S3, returns a dataframe based on the parameters provided.
}
