% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEventToS3.R
\name{sendEventToS3}
\alias{sendEventToS3}
\title{sendEventToS3(d, event_type, user_input)}
\usage{
sendEventToS3(d = NULL, event_type, user_input = NULL)
}
\arguments{
\item{d}{a list object that represents app information, timestamps etc}

\item{event_type}{a type of even such as LOGIN, LOGOUT OR READ_S3}

\item{user_input}{information about a user for analysis}
}
\description{
Standard function for connecting to S3.
}
