% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_list_bucket_items.R
\name{s3_list_bucket_items}
\alias{s3_list_bucket_items}
\title{s3_list_bucket_items(bucket)}
\usage{
s3_list_bucket_items(bucket)
}
\arguments{
\item{bucket}{your aws bucket}
}
\description{
Standard function for connecting to S3.
}
