% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_event_to_s3.R
\name{send_event_to_s3}
\alias{send_event_to_s3}
\title{send_event_to_s3(d, event_type, user_input)}
\usage{
send_event_to_s3(d = NULL, event_type, user_input = NULL)
}
\arguments{
\item{d}{a list object that represents app information, timestamps etc}

\item{event_type}{a type of even such as LOGIN, LOGOUT OR READ_S3}

\item{user_input}{information about a user for analysis}
}
\description{
Standard function for connecting to S3.
}
